+++ src/main/java/dao/InvitedDAO.java
@@ -5,7 +5,7 @@ import java.util.List;
 import models.Invited;
 
 public interface InvitedDAO {
-    long saveInvited(Invited invited);
+    int saveInvited(Invited invited);
     List<Invited> findAllInvitedUsers(long storyId);
     int removeInvited(Invited invited);
 }
